+++ Python Backend/dateUpdate.txt
@@ -4,21 +4,3 @@
 349,4900.000000,-0.179932,-0.015137,0.950439,28.271599,1001.191506,0.000000,0.000000
 350,4900.000000,-0.185303,-0.018066,0.938477,28.271599,1001.191506,0.000000,0.000000
 351,4902.500000,-0.180420,-0.019043,0.946289,28.288227,1001.113029,0.000000,0.000000
-352,4901.250000,-0.181396,-0.015625,0.941895,28.304855,1001.034543,0.000000,0.000000
-353,4900.000000,-0.178467,-0.019287,0.949707,28.304855,1001.154428,0.000000,0.000000
-354,4900.000000,-0.182617,-0.019043,0.945068,28.288227,1000.993148,0.000000,0.000000
-355,4901.250000,-0.179688,-0.018311,0.948242,28.288227,1001.113029,0.000000,0.000000
-356,4897.500000,-0.186768,-0.016846,0.946045,28.321484,1001.195828,0.000000,0.000000
-357,4900.000000,-0.173584,-0.020996,0.949951,28.304855,1001.394199,0.000000,0.000000
-358,4901.250000,-0.182617,-0.019775,0.955811,28.271599,1001.191506,0.000000,0.000000
-292,4900.000000,-0.178223,-0.020020,0.942139,28.105314,1001.256801,0.000000,0.000000
-293,4900.000000,-0.182617,-0.011719,0.941650,28.138571,1001.099959,0.000000,0.000000
-294,4901.250000,-0.177979,-0.016113,0.943359,28.138571,1001.219801,0.000000,0.000000
-295,4900.000000,-0.188721,-0.017578,0.950195,28.171828,1001.182784,0.000000,0.000000
-296,4901.250000,-0.179443,-0.016113,0.942871,28.138571,1001.099959,0.000000,0.000000
-297,4900.000000,-0.181885,-0.014404,0.937500,28.171828,1001.182784,0.000000,0.000000
-298,4900.000000,-0.181152,-0.010498,0.942871,28.188457,1001.104341,0.000000,0.000000
-299,4860.000000,-0.180176,-0.016846,0.943848,28.171828,1001.182784,0.000000,0.000000
-300,4898.750000,-0.176025,-0.017334,0.944580,28.121943,1001.178384,0.000000,0.000000
-301,4900.000000,-0.188721,-0.009521,0.947998,28.105314,1001.256801,0.000000,0.000000
-302,4900.000000,-0.176758,-0.019043,0.934326,28.072058,1001.173958,0.000000,0.000000
