+++ src/main/main.ts
@@ -127,9 +127,6 @@ const createWindow = async () => {
   // Remove this if your app does not use auto updates
   // eslint-disable-next-line
   new AppUpdater();
-
-
-  makeTestPostRequest();
 };
 
 /**
